// PI_4eme-Template/PI_4eme-Template/Platforme/src/app/admin/b2b/models/b2b.models.ts
export interface Company {
  id: number;
  name: string;
  email: string;
  siret: string;
  sector: string;
  address: string;
  phone: string;
  creditsRemaining: number;
  createdBy: number;
  createdAt: string;
  employeeCount: number;
}

export interface CompanyRequest {
  name: string;
  email: string;
  siret: string;
  sector: string;
  address: string;
  phone: string;
  createdBy: number;
}